+++ src/router/util.js
@@ -1,8 +1,0 @@
-'use strict';
-
-exports.fmt = function (name, val) {
-	if (name === 'url') {
-		val = decodeURI(val);
-	}
-	return '(' + name + ':' + val + ')';
-};