+++ CHANGELOG.md
@@ -2,6 +2,28 @@
 
 This is a list of manually mantained changes and updates for each version.
 
+## Version 2.5.0
+
+## Added
+
+None
+
+## Changed
+
+None
+
+## Deprecated
+
+#### gracenode.registerShutdownTask()
+
+#### gracenode.router
+
+Please use `gracenode.http` instead.
+
+## Removed
+
+None
+
 ## Version 2.4.1
 
 ## Added